last changes

+++ src/main/java/com/airportspolish/SRB/model/Temp.java
@@ -36,7 +36,7 @@ public class Temp {
     @Transient
     private boolean tempNewRecord;
     @Transient
-    private Long tempCloseTypeId;
+    private int tempCloseTypeId;
